python-k8s-manager/requirements.txt
Flask==3.0.0
kubernetes==28.1.0
nats-py==2.6.0
flask-cors==4.0.0
